vec_deque tests: remove static mut
